+++ src/mprPath.c
@@ -948,7 +948,7 @@ char *mprGetNormalizedPath(MprCtx ctx, cchar *pathArg)
      *  Allocate one spare byte incase we need to break into segments. If so, will add a trailing
      *  "/" to make parsing easier later.
      */
-    len = strlen(pathArg);
+    len = (int) strlen(pathArg);
     if ((path = mprAlloc(ctx, len + 2)) == 0) {
         return NULL;
     }
@@ -1001,7 +1001,7 @@ char *mprGetNormalizedPath(MprCtx ctx, cchar *pathArg)
      *  Have dots to process so break into path segments. Add one incase we need have an absolute path with a drive-spec.
      */
     mprAssert(segmentCount > 0);
-    if ((segments = mprAlloc(ctx, sizeof(char*) * (segmentCount + 1))) == 0) {
+    if ((segments = mprAlloc(ctx, (int) sizeof(char*) * (segmentCount + 1))) == 0) {
         mprFree(dupPath);
         return NULL;
     }
@@ -1034,7 +1034,7 @@ char *mprGetNormalizedPath(MprCtx ctx, cchar *pathArg)
                 continue;
             }
             segments[i++] = mark;
-            len += sp - mark;
+            len += (int) (sp - mark);
 #if KEEP
             if (i == 1 && segmentCount == 1 && fs->hasDriveSpecs && strchr(mark, ':') != 0) {
                 /*
@@ -1050,7 +1050,7 @@ char *mprGetNormalizedPath(MprCtx ctx, cchar *pathArg)
 
     if (--sp > mark) {
         segments[i++] = mark;
-        len += sp - mark;
+        len += (int) (sp - mark);
     }
     mprAssert(i <= segmentCount);
     segmentCount = i;
@@ -1464,7 +1464,7 @@ char *mprGetAppPath(MprCtx ctx)
         return mprGetAbsPath(ctx, ".");
     }
     path[size] = '\0';
-    len = readlink(path, pbuf, sizeof(pbuf) - 1);
+    len = (int) readlink(path, pbuf, sizeof(pbuf) - 1);
     if (len < 0) {
         return mprGetAbsPath(ctx, path);
     }
