+++ src/regexp/pcre_valid_utf8.c
@@ -83,7 +83,7 @@ register const uschar *p;
 if (length < 0)
   {
   for (p = string; *p != 0; p++);
-  length = p - string;
+  length = (int) (p - string);
   }
 
 for (p = string; length-- > 0; p++)
@@ -91,13 +91,13 @@ for (p = string; length-- > 0; p++)
   register int ab;
   register int c = *p;
   if (c < 128) continue;
-  if (c < 0xc0) return p - string;
+  if (c < 0xc0) return (int) (p - string);
   ab = _pcre_utf8_table4[c & 0x3f];     /* Number of additional bytes */
-  if (length < ab || ab > 3) return p - string;
+  if (length < ab || ab > 3) return (int) (p - string);
   length -= ab;
 
   /* Check top bits in the second byte */
-  if ((*(++p) & 0xc0) != 0x80) return p - string;
+  if ((*(++p) & 0xc0) != 0x80) return (int) (p - string);
 
   /* Check for overlong sequences for each different length, and for the
   excluded range 0xd000 to 0xdfff.  */
@@ -107,7 +107,7 @@ for (p = string; length-- > 0; p++)
     /* Check for xx00 000x (overlong sequence) */
 
     case 1:
-    if ((c & 0x3e) == 0) return p - string;
+    if ((c & 0x3e) == 0) return (int) (p - string);
     continue;   /* We know there aren't any more bytes to check */
 
     /* Check for 1110 0000, xx0x xxxx (overlong sequence) or
@@ -116,7 +116,7 @@ for (p = string; length-- > 0; p++)
     case 2:
     if ((c == 0xe0 && (*p & 0x20) == 0) ||
         (c == 0xed && *p >= 0xa0))
-      return p - string;
+      return (int) (p - string);
     break;
 
     /* Check for 1111 0000, xx00 xxxx (overlong sequence) or
@@ -126,7 +126,7 @@ for (p = string; length-- > 0; p++)
     if ((c == 0xf0 && (*p & 0x30) == 0) ||
         (c > 0xf4 ) ||
         (c == 0xf4 && *p > 0x8f))
-      return p - string;
+      return (int) (p - string);
     break;
 
 #if 0
@@ -151,7 +151,7 @@ for (p = string; length-- > 0; p++)
   /* Check for valid bytes after the 2nd, if any; all must start 10 */
   while (--ab > 0)
     {
-    if ((*(++p) & 0xc0) != 0x80) return p - string;
+    if ((*(++p) & 0xc0) != 0x80) return (int) (p - string);
     }
   }
 #endif
